+++ app/api/v1/health.rb
@@ -1,6 +1,0 @@
-class V1::Health < Grape::API
-  desc "Returns 'ok' if API is running properly", tags: ['health']
-  get :health do
-    { status: 'ok' }
-  end
-end