+++ app/assets/javascripts/controllers/beta/project_boards/ShowController.jsx
@@ -1,0 +1,17 @@
+import React from 'react';
+import { render } from 'react-dom';
+import { Provider } from 'react-redux';
+import ProjectBoard from 'components/projects/ProjectBoard';
+import store from 'store';
+
+export default () => {
+  const appElement = document.querySelector('[data-app]');
+  const { projectId } = appElement.dataset;
+
+  render(
+    <Provider store={store}>
+      <ProjectBoard projectId={projectId} />
+    </Provider>,
+    appElement
+  );
+};