+++ app/assets/javascripts/libs/execute_attachinary.js
@@ -1,5 +1,7 @@
 const ExecuteAttachinary = () => {
-  $('.attachinary-input').attachinary({ template: $('#attachinary_template').html() });
+  $('.attachinary-input').attachinary({
+    template: $('#attachinary_template').html(),
+  });
 };
 
 export default ExecuteAttachinary;